Update to Script
Added 
Sentiment analysis with TextBlob
In this version of the function, the TextBlob library is used to classify the sentiment of each tweet. The resulting polarity value is stored in the sentiment column of the tweets table along with the tweet text. The sentiment column should be of a floating-point type to store decimal values.

Added
the ssc.awaitTerminationOrTimeout(timeout) method to specify a timeout, after which the streaming context will be stopped if it has not already terminated.